MNT: Move metar_parser.py up a level
Replace private subpackage with a private module since we moved
everything else out of that private directory.
