+++ setup.cfg
@@ -9,7 +9,7 @@ copyright-author = MetPy Developers
 rst-roles = class, data, doc, func, meth, mod
 rst-directives = plot, versionchanged
 known-modules = matplotlib:[matplotlib,mpl_toolkits],netcdf4:[netCDF4]
-exclude = docs build src/metpy/io/_metar_parser/metar_parser.py
+exclude = docs build src/metpy/io/_metar_parser.py
 select = E301 E302 E303 E304 E305 E306 I N400 R
 ignore = F405 W503 RST902 SIM106
 per-file-ignores = examples/*.py: D MPY001
